Eliminate MSVC warnings

+++ DemandLoading/TexturePainting/CanvasImage.cpp
@@ -158,7 +158,7 @@ void CanvasImage::drawBrush( CanvasBrush& brush, int xcenter, int ycenter )
 
 void CanvasImage::drawStroke( CanvasBrush& brush, int x0, int y0, int x1, int y1 )
 {
-    float d = sqrtf( (x1-x0)*(x1-x0) + (y1-y0)*(y1-y0) );
+    float d = sqrtf( static_cast<float>( (x1-x0)*(x1-x0) + (y1-y0)*(y1-y0) ) );
     int numIterations = std::max( static_cast<int>( 10 * d / brush.m_width ), 1 );
     for( int i=0; i<numIterations; ++i )
     {
